Add vhost creation
